Colorize timestamp
