+++ create-reporter/index.test.ts
@@ -329,20 +329,20 @@ it('reports denied', async () => {
 it('reports unknownType', async () => {
   await check('unknownType', {
     type: 'CHANGE_SER',
-    actionId: '1487805099387 100:uImkcF4z 0'
+    actionId: '1487805099387 100:vAApgNT9 0'
   })
 })
 
 it('reports unknownType from server', async () => {
   await check('unknownType', {
     type: 'CHANGE_SER',
-    actionId: '1487805099387 server:FnXaqDxY 0'
+    actionId: '1650269021700 server:FnXaqDxY 0'
   })
 })
 
 it('reports wrongChannel', async () => {
   await check('wrongChannel', {
-    actionId: '1487805099387 100:uImkcF4z 0',
+    actionId: '1650269045800 100:IsvVzqWx 0',
     channel: 'ser/100'
   })
 })
@@ -350,7 +350,7 @@ it('reports wrongChannel', async () => {
 it('reports wrongChannel without name', async () => {
   await check('wrongChannel', {
     channel: undefined,
-    actionId: '1487805099387 100:uImkcF4z 0'
+    actionId: '1650269056600 100:uImkcF4z 0'
   })
 })
